requirements.txt
fastapi==0.100.1
uvicorn

psycopg2-binary
sqlalchemy==1.4.27

pytest==7.1.2
pytest-asyncio==0.18.3
python-multipart==0.0.5
coverage

python-dotenv
requests
httpx